s3_worker/requirements.txt
azure-storage-blob>=12.20.0
pymysql>=1.1.0
cryptography>=44.0.0
python-dotenv>=1.0.0
